Added whitelist + ban database struct
Update it. Do it now.
By default, when creating an account, it's not banned and NOT
whitelisted.
If you are updating your dbstruct and want to whitelist all accounts,
use this query after updating the struct:
"UPDATE rotmg.accounts SET whitelisted='1'" - without the doublequotes
(from MySQL Workbench - you may have to disable the safe mode)